requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
nfl_data_py>=0.3.0
pytrends>=4.9.0
python-dotenv>=1.0.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
matplotlib>=3.7.0
